PMM-6091 Rename UpdateModalHeader to ProgressModalHeader
