+++ pmm-app/src/pmm-update/components/ProgressModalHeader/ProgressModalHeader.test.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { UpdateModalHeader } from './UpdateModalHeader';
-import { Messages } from './UpdateModalHeader.messages';
+import { ProgressModalHeader } from './ProgressModalHeader';
+import { Messages } from './ProgressModalHeader.messages';
 
-describe('UpdateModalHeader::', () => {
+describe('ProgressModalHeader::', () => {
   it('should show that the update is in progress by default', () => {
-    const wrapper = shallow(<UpdateModalHeader />);
+    const wrapper = shallow(<ProgressModalHeader />);
 
     expect(wrapper.find('h4').text()).toEqual(Messages.updateInProgress);
 
@@ -14,7 +14,7 @@ describe('UpdateModalHeader::', () => {
   });
 
   it('should show that the update succeeded if isUpdated is true', () => {
-    const wrapper = shallow(<UpdateModalHeader isUpdated />);
+    const wrapper = shallow(<ProgressModalHeader isUpdated />);
 
     expect(wrapper.find('h4').text()).toEqual(Messages.updateSucceeded);
 
@@ -22,7 +22,7 @@ describe('UpdateModalHeader::', () => {
   });
 
   it('should show ignore updateFailed if isUpdated is true', () => {
-    const wrapper = shallow(<UpdateModalHeader isUpdated updateFailed />);
+    const wrapper = shallow(<ProgressModalHeader isUpdated updateFailed />);
 
     expect(wrapper.find('h4').text()).toEqual(Messages.updateSucceeded);
 
@@ -31,7 +31,7 @@ describe('UpdateModalHeader::', () => {
 
   it('should show the passed error message if the update failed', () => {
     const errorMessage = 'Test Error';
-    const wrapper = shallow(<UpdateModalHeader updateFailed errorMessage={errorMessage} />);
+    const wrapper = shallow(<ProgressModalHeader updateFailed errorMessage={errorMessage} />);
 
     expect(
       wrapper
